Adding transactional functionality

+++ challenge/src/main/java/com/dws/challenge/domain/AmountDto.java
@@ -1,5 +1,7 @@
 package com.dws.challenge.domain;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,13 +14,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AmountDto {
-    private double amount;
+	
+    private BigDecimal amount;
 
-	public double getAmount() {
+	public BigDecimal getAmount() {
 		return amount;
 	}
 
-	public void setAmount(double amount) {
+	public void setAmount(BigDecimal amount) {
 		this.amount = amount;
 	}
 }
